feat: family configuration
This is a configuration for the family laptop.

As usual, start with huge copy paste and later we'll iteratively share /
improve.

Note that for the family computer, I tried an approach without password,
so enabled `fprintd` and actually tried the same on my setup because it
seems to work great.